requirements.txt
Flask==2.3.3
Authlib==1.3.0
python-dotenv==1.0.0
cryptography==41.0.7
garminconnect==0.1.50
rq==1.15.1
redis==5.0.1 